FIX: fix bit fields for mingw/windows builds
It seems like it inserts padding when there's space between the current
field and the next one, if they're differing types.

e.g. uint8_t a : 3; uint16_t b : 5; would produce a 2-byte struct even
though it looks like the bits should be packed. GCC and clang on Linux
handle this as expected.